+++ packages/adminui-vue/utils/Form/Form.js
@@ -20,11 +20,11 @@ const UB = require('@unitybase/ub-pub')
  * Creates a new instance of UI module what can be visualized in the application shell using `mount*()` call
  * @param {object} cfg
  * @param {Vue.Component} [cfg.component] Form component
- * @param {Vue.Component} [cfg.rootComponent] Alias for cfg.component
+ * @param {Vue.Component} [cfg.rootComponent] Alias for `cfg.component`
  * @param {object} [cfg.props] Form component props
  * @param {object} [cfg.props.parentContext] Attributes values what will be passed to addNew method
  *   in case instanceID is empty. Think of it as default values for attributes of a new record
- * @param {string} [cfg.title] Form title
+ * @param {string} [cfg.title] Form title. Can contain macros `{attrName}`, such macros will be replaced by attributes values. Example - `({code}) {name}`
  * @param {string} cfg.entity Entity name for master record
  * @param {number} [cfg.instanceID] Instance ID
  * @param {boolean} [cfg.isModal=false] If true form will be displayed inside modal dialog. Otherwise - in tab (default)