+++ packages/cdn/public/forms/cdn_region-1-fm.vue
@@ -30,7 +30,10 @@ const { Form } = require('@unitybase/adminui-vue')
 const { mapGetters } = require('vuex')
 
 module.exports.mount = cfg => {
-  Form(cfg)
+  Form({
+    ...cfg,
+    title: '{code} {name}'
+  })
     .processing()
     .validation()
     .mount()