+++ packages/org/public/forms/org_unit-fm.vue
@@ -19,7 +19,10 @@ const { Form, mapInstanceFields } = require('@unitybase/adminui-vue')
 const { mapGetters } = require('vuex')
 
 module.exports.mount = cfg => {
-  Form(cfg)
+  Form({
+    ...cfg,
+    title: '{code} {name}'
+  })
     .processing()
     .validation()
     .mount()