+++ packages/cdn/public/forms/cdn_person-fm.vue
@@ -97,7 +97,10 @@ const { Form, mapInstanceFields } = require('@unitybase/adminui-vue')
 const { mapState, mapGetters } = require('vuex')
 
 module.exports.mount = cfg => {
-  Form(cfg)
+  Form({
+    ...cfg,
+    title: '{firstName} {lastName}'
+  })
     .processing()
     .validation()
     .mount()