+++ packages/org/public/forms/org_department-fm.vue
@@ -51,7 +51,10 @@ const { Form, mapInstanceFields } = require('@unitybase/adminui-vue')
 const { mapState, mapGetters } = require('vuex')
 
 module.exports.mount = cfg => {
-  Form(cfg)
+  Form({
+    ...cfg,
+    title: '{code} {name}'
+  })
     .processing({
       masterFieldList: ['parentID', 'depTypeID', 'code', 'name', 'fullName', 'description', 'isClerical',
         'nameGen', 'fullNameGen', 'nameDat', 'fullNameDat']